Удален класс Url, формирование ссылок передано  классу WebRouter, который наследуется от Router и заменяет его.
